refactor: changed roomService import
