+++ src/app/(main)/manage/(dashboard)/page.tsx
@@ -6,7 +6,7 @@ export const metadata: Metadata = {
   title: "Manager Dashboard",
 };
 import { housingService } from "@/app/lib/services/housing-service";
-import { roomService } from "@/app/lib/services/room-service";
+import * as roomService from "@/app/lib/services/room-service";
 import Link from "next/link";
